chore(): move schema to /prisma
